+++ app/components/layout/settings/index.ts
@@ -1,3 +1,2 @@
 export * from "./settings-button";
 export * from "./settings-modal";
-export { default as SettingsUpdates } from "./updates";